+++ src/textual/widgets/_tree.py
@@ -464,56 +464,100 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
 
         Can be handled using `on_tree_node_collapsed` in a subclass of `Tree` or in a
         parent node in the DOM.
-
-        Attributes:
-            node: The node that was collapsed.
         """
 
-        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
+        def __init__(
+            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
+        ) -> None:
+            self.tree = tree
+            """The tree that sent the message."""
             self.node: TreeNode[EventTreeDataType] = node
+            """The node that was collapsed."""
             super().__init__()
+
+        @property
+        def control(self) -> Tree[EventTreeDataType]:
+            """The tree that sent the message.
+
+            This is an alias for [`NodeCollapsed.tree`][textual.widgets.Tree.NodeCollapsed.tree]
+            and is used by the [`on`][textual.on] decorator.
+            """
+            return self.tree
 
     class NodeExpanded(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is expanded.
 
         Can be handled using `on_tree_node_expanded` in a subclass of `Tree` or in a
         parent node in the DOM.
-
-        Attributes:
-            node: The node that was expanded.
         """
 
-        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
+        def __init__(
+            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
+        ) -> None:
+            self.tree = tree
+            """The tree that sent the message."""
             self.node: TreeNode[EventTreeDataType] = node
+            """The node that was expanded."""
             super().__init__()
+
+        @property
+        def control(self) -> Tree[EventTreeDataType]:
+            """The tree that sent the message.
+
+            This is an alias for [`NodeExpanded.tree`][textual.widgets.Tree.NodeExpanded.tree]
+            and is used by the [`on`][textual.on] decorator.
+            """
+            return self.tree
 
     class NodeHighlighted(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is highlighted.
 
         Can be handled using `on_tree_node_highlighted` in a subclass of `Tree` or in a
         parent node in the DOM.
-
-        Attributes:
-            node: The node that was highlighted.
         """
 
-        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
+        def __init__(
+            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
+        ) -> None:
+            self.tree = tree
+            """The tree that sent the message."""
             self.node: TreeNode[EventTreeDataType] = node
+            """The node that was highlighted."""
             super().__init__()
+
+        @property
+        def control(self) -> Tree[EventTreeDataType]:
+            """The tree that sent the message.
+
+            This is an alias for [`NodeHighlighted.tree`][textual.widgets.Tree.NodeHighlighted.tree]
+            and is used by the [`on`][textual.on] decorator.
+            """
+            return self.tree
 
     class NodeSelected(Generic[EventTreeDataType], Message, bubble=True):
         """Event sent when a node is selected.
 
         Can be handled using `on_tree_node_selected` in a subclass of `Tree` or in a
         parent node in the DOM.
-
-        Attributes:
-            node: The node that was selected.
         """
 
-        def __init__(self, node: TreeNode[EventTreeDataType]) -> None:
+        def __init__(
+            self, tree: Tree[EventTreeDataType], node: TreeNode[EventTreeDataType]
+        ) -> None:
+            self.tree = tree
+            """The tree that sent the message."""
             self.node: TreeNode[EventTreeDataType] = node
+            """The node that was selected."""
             super().__init__()
+
+        @property
+        def control(self) -> Tree[EventTreeDataType]:
+            """The tree that sent the message.
+
+            This is an alias for [`NodeSelected.tree`][textual.widgets.Tree.NodeSelected.tree]
+            and is used by the [`on`][textual.on] decorator.
+            """
+            return self.tree
 
     def __init__(
         self,
@@ -813,7 +857,7 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
             node._selected = True
             self._cursor_node = node
             if previous_node != node:
-                self.post_message(self.NodeHighlighted(node))
+                self.post_message(self.NodeHighlighted(self, node))
 
     def watch_guide_depth(self, guide_depth: int) -> None:
         self._invalidate()
@@ -1065,10 +1109,10 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
             return
         if node.is_expanded:
             node.collapse()
-            self.post_message(self.NodeCollapsed(node))
+            self.post_message(self.NodeCollapsed(self, node))
         else:
             node.expand()
-            self.post_message(self.NodeExpanded(node))
+            self.post_message(self.NodeExpanded(self, node))
 
     async def _on_click(self, event: events.Click) -> None:
         meta = event.style.meta
@@ -1155,4 +1199,4 @@ class Tree(Generic[TreeDataType], ScrollView, can_focus=True):
             node = line.path[-1]
             if self.auto_expand:
                 self._toggle_node(node)
-            self.post_message(self.NodeSelected(node))
+            self.post_message(self.NodeSelected(self, node))